# swagger/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
  /rest/file/{content}:
    post:
      tags:
      - rest-api
      operationId: file
      parameters:
      - name: content
        in: path
        required: true
        schema:
          type: string
          maxLength: 1000
          minLength: 0
      responses:
        "200":
          description: OK
  /orders:
    get:
      tags:
      - order-api
      operationId: orders
      parameters:
      - name: customer
        in: query
        required: false
        schema:
          type: string
      - name: city
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      tags:
      - order-api
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Order"
  /countries:
    post:
      tags:
      - banking-api
      operationId: createCountry
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCountryDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CountryDto"
  /accounts:
    get:
      tags:
      - banking-api
      operationId: accounts
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountDto"
    post:
      tags:
      - banking-api
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountDto"
  /rest/risk:
    get:
      tags:
      - rest-api
      operationId: risk
      responses:
        "200":
          description: OK
  /rest/posts:
    get:
      tags:
      - rest-api
      operationId: posts
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostDto"
  /rest/passphrase:
    get:
      tags:
      - rest-api
      operationId: passphrase
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /rest/cookies:
    get:
      tags:
      - rest-api
      operationId: cookies
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
  /cards:
    get:
      tags:
      - banking-api
      operationId: cards
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardDto"
components:
  schemas:
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        postalCode:
          type: string
    Item:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        quantity:
          type: integer
          format: int32
        price:
          type: number
    Order:
      type: object
      properties:
        id:
          type: string
        customer:
          type: string
        address:
          $ref: "#/components/schemas/Address"
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        total:
          type: number
        createdAt:
          type: string
          format: date-time
    CreateCountryDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 0
        code:
          type: string
          maxLength: 10
          minLength: 0
          pattern: "[A-Za-z0-9]+"
      required:
      - code
      - name
    CountryDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 50
          minLength: 0
        code:
          type: string
          maxLength: 10
          minLength: 0
          pattern: "[A-Za-z0-9]+"
        insertTime:
          type: string
          format: date-time
        version:
          type: integer
          format: int32
      required:
      - code
      - id
      - name
      - version
    CreateAccountDto:
      type: object
      properties:
        accountNumber:
          type: string
          minLength: 1
          pattern: "\\d{26}"
      required:
      - accountNumber
    AccountDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        accountNumber:
          type: string
          minLength: 1
          pattern: "\\d{26}"
      required:
      - accountNumber
      - id
      - version
    PostDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        title:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
      required:
      - body
      - id
      - title
      - userId
    CardDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        cardNumber:
          type: string
          minLength: 1
          pattern: "\\d{16}"
        accountNumber:
          type: string
          pattern: "\\d{26}"
      required:
      - cardNumber
      - id
      - version